refactor(desktop): extract shared Pierre-tree module + decompose FilesTab
- New renderer/lib/pierreTree: createPierreTreeStyle (the ~55-line --trees-*
  CSS-var map, parameterized by row height / indent / search chrome) and the
  FILE_STATUS_TO_PIERRE git-status mapping — both were duplicated verbatim in
  FilesTab and ChangesTreeView.
- FilesTab (722 → ~290 lines): pull the filesystem actions (create / rename /
  delete / reveal / collapse-all + the bridge bookkeeping) into a
  useFilesTabActions hook; move buildPierreGitStatus and the creation-path
  helpers into utils/; lift the header icon button into its own component;
  add a local constants.ts (FILE_EXPLORER_ROW_HEIGHT/INDENT/OVERSCAN), which
  also drops FilesTab's last import from the sunsetting v1 FilesView code.
- ChangesTreeView: use the shared style + status map; extract the content-
  height measurement into a useMeasuredTreeHeight hook and buildTreeShape into
  utils/.

No behavior change.
