+++ apps/desktop/src/renderer/lib/pierreTree/index.ts
@@ -1,0 +1,7 @@
+export { createPierreTreeStyle } from "./createPierreTreeStyle";
+export {
+	FILE_STATUS_TO_PIERRE,
+	type FileStatus,
+	type PierreGitStatus,
+	type PierreGitStatusEntry,
+} from "./pierreGitStatus";
