+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/components/WorkspaceSidebar/components/FilesTab/components/FilesTabHeaderButton/index.ts
@@ -1,0 +1,1 @@
+export { FilesTabHeaderButton } from "./FilesTabHeaderButton";
