+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/components/WorkspaceSidebar/components/FilesTab/constants.ts
@@ -1,0 +1,6 @@
+/** Row height (px) for the Files-tab explorer tree — drives the Pierre model's `itemHeight` and the `--trees-row-height-override`. */
+export const FILE_EXPLORER_ROW_HEIGHT = 28;
+/** Per-level indent (px) for the Files-tab explorer tree. */
+export const FILE_EXPLORER_INDENT = 10;
+/** Rows rendered beyond the viewport in the virtualized explorer tree. */
+export const FILE_EXPLORER_OVERSCAN = 10;
