+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/components/WorkspaceSidebar/components/FilesTab/utils/buildPierreGitStatus/index.ts
@@ -1,0 +1,1 @@
+export { buildPierreGitStatus } from "./buildPierreGitStatus";
