+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/components/WorkspaceSidebar/hooks/useChangesTab/components/ChangesFileList/components/ChangesTreeView/utils/buildTreeShape/index.ts
@@ -1,0 +1,1 @@
+export { buildTreeShape, type TreeShape } from "./buildTreeShape";
